requirements.txt
eyed3==0.9.7
ffmpy==0.3.0
Flask==2.2.2
Pillow==9.4.0
pytube==12.1.2
requests==2.28.1
setuptools==58.1.0
